Rename packages of Interop Cordapps to adhere to main.
