+++ components/interop/interop-service/build.gradle
@@ -34,7 +34,7 @@ dependencies {
 
     testImplementation "org.assertj:assertj-core:$assertjVersion"
     testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
-    testImplementation project(":libs:flows:utils")
+    testImplementation project(":libs:flows:flow-utils")
 
     integrationTestImplementation project(":testing:db-message-bus-testkit")
