+++ gradle.properties
@@ -44,7 +44,7 @@ bouncycastleVersion=1.73
 # Corda API libs revision (change in 4th digit indicates a breaking change)
 # Change to 5.0.0.xx-SNAPSHOT to pick up maven local published copy
 #cordaApiVersion=5.0.0.xxx-SNAPSHOT
-cordaApiVersion=5.0.0-INTEROP.7-beta+
+cordaApiVersion=5.0.0-INTEROP.8-alpha-1683030060796
 
 disruptorVersion=3.4.4
 felixConfigAdminVersion=1.9.26
