+++ testing/cpbs/test-invocation-cordapp/src/main/kotlin/com/r3/corda/testing/interop/FacadeInvocationFlow.kt
@@ -1,4 +1,4 @@
-package net.cordapp.testing.testflows
+package com.r3.corda.testing.interop
 
 import net.corda.v5.application.flows.ClientStartableFlow
 import net.corda.v5.application.flows.CordaInject
